Test For Interpolation
We do a single step with length specified by step_len. The
approximations are done by calling  the interpolate() function of the
stepper. The approximations at different fractions (tau)  of the step,
as specified in the tau_list The resulting values are compared with a
ref. stepper (like G4CashKarpRKF45) with repeated 	steps with reduces
step sizes (step_len*tau)